refactor(inspect): wrap the board charts in a component (no raw ECharts in the view)
InspectView instantiated ECharts directly, violating the "charts are wrapped" rule. The shared TimeChart is line/area-only, but the inspect board also has a bar mode, so extract a feature-local InspectMetricChart.vue that owns one ECharts instance + its lifecycle (a ResizeObserver replaces the manual window-resize + density-watch wiring; Vue's per-widget mount/unmount replaces the manual instance Map and disposal). Behavior is unchanged — the chart option builder moved verbatim. Net -160 lines in the view.
